% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipaint.R
\name{ipaint}
\alias{ipaint}
\title{Interatively scroll through rows of a painted dataframe}
\usage{
ipaint(df = .Last.value)
}
\arguments{
\item{df}{the dataframe to scroll through, defaults to \code{.Last.value}}
}
\description{
This function calls \code{\link[=paint]{paint()}} repeatedly on a dataframe,
updating the rows shown in response to left or right keypresses.
}
\details{
It will only work in terminals supported by {keypress} - Not many!
}
