% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_print.R
\name{mask_print}
\alias{mask_print}
\alias{unmask_print}
\title{mask print methods for paintable dataframes}
\usage{
mask_print()

unmask_print()
}
\description{
This function replaces the original paint methods for supported data.frame
sub classes with calls to \code{paint()} in the current session. The utility of
this is that \code{paint()} will be called whenever \code{print()} would, for example
at the end of an expression when coding interactively.
}
\details{
This also makes paint usable with other tools that output to console e.g. \code{{breakerofchains}}.

If you really like \code{paint()} you could put \code{paint::mask_print()} in your \code{.Rprofile}.

By default a colourful reminder message is emitted whenever print methods are
replaced, which may not happen until the package that contains them is actually
loaded, e.g. after \code{library(tibble)}. This can be disabled with \code{options(paint_remind_mask_print = FALSE)}
}
\section{Functions}{
\itemize{
\item \code{unmask_print}: the reverse operation to \code{mask_print}: return all print methods to defaults.
}}

\seealso{
\code{\link[=unpaint]{unpaint()}} A method for calling the default print functions once for a single object or .Last.value.
}
