% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint.R
\name{paint}
\alias{paint}
\title{Paint a rectangle in the data.frame family}
\usage{
paint(df, ...)
}
\arguments{
\item{df}{the dataframe to paint}

\item{...}{further arguments to class specific paint methods}
}
\value{
df invisibly. Text outputed printed to console.
}
\description{
{paint} is an alternative print method of objects that are of class
\code{data.frame}. It has been described as a "pretty \code{\link[pillar:glimpse]{pillar::glimpse()}}", since it
also favours printing columns running horizontally.
}
\details{
{paint} tries to be low noise by using colour instead of markup characters.
It also highlights important values like \code{NA}, \code{NULL}, \code{NaN}, and \code{Inf}.

Each supported sub-class of \code{data.frame} has a custom meta data header,
designed to show useful information for wrangling.

The appearance of the output can be customised with options:
\itemize{
\item \code{paint_palette} the palette to paint row colours with, defaults to
\code{rainbow_6()}, 6 standard ANSI colours that are supported in most terminals.
\item \code{paint_n_rows} the number of rows to print for each column. Defaults to the
number of colours in the \code{paint_palette}, but can be set higher for repeating sequences.
\item \code{paint_max_width} the maximum width of the output. 60 is the default. \code{{paint}} does not resize
based on terminal width by design.
\item \code{paint_align_row_head = c("left", "center", "right")}. How to align the column
title and type. Defaults to \code{"left"}.
\item \code{paint_dark_mode} darken the \code{paint_palette} using \code{crayon::blurred} - not
supported in all terminals.
}
}
\examples{
paint(mtcars)
}
\seealso{
\code{\link[=unpaint]{unpaint()}} for calling the default print method on a dataframe
}
