% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint_stub.R
\name{paint_stub}
\alias{paint_stub}
\title{stub}
\usage{
paint_stub(object)
}
\description{
The stub is the bit after the name in the object title e.g. tibble \link[=row,
cols]{row, cols} - stub is the dimensions summary after "tibble"
}
