% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rainbow_text}
\alias{rainbow_text}
\title{Colour some text with a {paint} palette}
\usage{
rainbow_text(msg)
}
\arguments{
\item{msg}{a length one character vector (not length checked) to be coloured
using the palette set in the \code{paint_palette} option.}
}
\value{
a character vector with ANSI colour codes embedded. use cat() to see the coloured message.
}
\description{
Exported just for fun.
}
