% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitise_col.R
\name{sanitise_col}
\alias{sanitise_col}
\title{Attempt to transform chars in encodings we don't understand, rather than crash crayon.}
\usage{
sanitise_col(col)
}
\arguments{
\item{col}{the column to sanitise, a vector}
}
\description{
Attempt to transform chars in encodings we don't understand, rather than crash crayon.
}
\author{
Miles McBain
}
