% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpaint.R
\name{unpaint}
\alias{unpaint}
\title{Call the default print methods instead of paint once}
\usage{
unpaint(df = .Last.value)
}
\arguments{
\item{df}{the object to print.}
}
\description{
If you're using \code{paint()} as your default \code{print()} for dataframes, you may
occasionally want to see an object printed using it's default method. This
function does that.
}
\details{
Either supply the object as the \code{df} arg, or if it has just been painted, you
can call \code{unpaint()} with no arguments to print the \code{.Last.value}.
}
\seealso{
\code{\link[=mask_print]{mask_print()}} to use \code{\link[=paint]{paint()}} by default for dataframes in the current session over \code{print()}.
}
