% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{rainbow_6}
\alias{rainbow_6}
\alias{viridis_6}
\alias{brewer_set3_12}
\alias{brewer_pastel1_8}
\alias{brewer_pastel2_7}
\alias{brewer_dark2_7}
\alias{brewer_accent_7}
\title{paintpals}
\usage{
rainbow_6()

viridis_6()

brewer_set3_12()

brewer_pastel1_8()

brewer_pastel2_7()

brewer_dark2_7()

brewer_accent_7()
}
\description{
paintpals
}
